Template: vite-cf-DO-v2-runner
Base template for Create an app for a construction contractor who charges a percentage of the BOM as his fees that will enable him enter detailed expenses for projects and then share with clients. Clients deposit a lumpsum then they can see utilisation and maybe pictures of progress. The contractor should be able to also manage his ledger and see maybe balancesheet and profit and loss if poasible. I think if he's also able to manage inventory of tools and sub-contractors and their specialisation that can be welcome. Eventually, we should be able to manage his suppliers and their costs so that he can be able to give a quick quote when clients ask for changes so that clients can decide quickly whether to proceed or not. Please do not use typescript and consider and orm